+++ test/tile_map/tile_test.dart
@@ -1,0 +1,15 @@
+import 'dart:math';
+
+import 'package:test/test.dart';
+import 'package:ziggurat/ziggurat.dart';
+
+void main() {
+  group('Tile', () {
+    test('Initialisation', () {
+      final tile = Tile(coordinates: Point(0, 0));
+      expect(tile.coordinates, equals(Point(0, 0)));
+      expect(tile.onEnter, isNull);
+      expect(tile.onExit, isNull);
+    });
+  });
+}